Move tests for scan results storage to an abstract class
This allows to reuse the same tests for different storage backends.
